+++ eras/alonzo/impl/src/Cardano/Ledger/Alonzo/TxCert.hs
@@ -21,11 +21,6 @@ instance Crypto c => EraTxCert (AlonzoEra c) where
   getTxCertPool (ShelleyTxCertPool c) = Just c
   getTxCertPool _ = Nothing
 
-  mkTxCertGenesis = ShelleyTxCertGenesis
-
-  getTxCertGenesis (ShelleyTxCertGenesis c) = Just c
-  getTxCertGenesis _ = Nothing
-
 instance Crypto c => ShelleyEraTxCert (AlonzoEra c) where
   {-# SPECIALIZE instance ShelleyEraTxCert (AlonzoEra StandardCrypto) #-}
 
@@ -33,6 +28,11 @@ instance Crypto c => ShelleyEraTxCert (AlonzoEra c) where
 
   getShelleyTxCertDeleg (ShelleyTxCertDelegCert c) = Just c
   getShelleyTxCertDeleg _ = Nothing
+
+  mkTxCertGenesisDeleg = ShelleyTxCertGenesisDeleg
+
+  getTxCertGenesisDeleg (ShelleyTxCertGenesisDeleg c) = Just c
+  getTxCertGenesisDeleg _ = Nothing
 
   mkTxCertMir = ShelleyTxCertMir