+++ eras/shelley/test-suite/test/Test/Cardano/Ledger/Shelley/Serialisation/Golden/Encoding.hs
@@ -111,7 +111,9 @@ import Cardano.Ledger.Shelley.TxBody (
   pattern RewardAcnt,
  )
 import Cardano.Ledger.Shelley.TxCert (
+  GenesisDelegCert (..),
   pattern ShelleyTxCertDeleg,
+  pattern TxCertGenesisDeleg,
   pattern TxCertMir,
  )
 import Cardano.Ledger.Shelley.TxWits (ShelleyTxWits, addrWits, scriptWits)
@@ -622,8 +624,8 @@ tests =
     , checkEncodingCBOR
         shelleyProtVer
         "genesis_delegation"
-        ( TxCertGenesis @C
-            ( ConstitutionalDelegCert @C_Crypto
+        ( TxCertGenesisDeleg @C
+            ( GenesisDelegCert @C_Crypto
                 testGKeyHash
                 (hashKey . vKey $ testGenesisDelegateKey @C_Crypto)
                 (testVRFKH @C_Crypto)