+++ libs/cardano-ledger-core/src/Cardano/Ledger/Core/TxCert.hs
@@ -11,14 +11,10 @@
 module Cardano.Ledger.Core.TxCert (
   EraTxCert (..),
   pattern TxCertPool,
-  pattern TxCertGenesis,
   Delegation (..),
   PoolCert (..),
   poolCWitness,
   poolCertKeyHashWitness,
-  ConstitutionalDelegCert (..),
-  genesisKeyHashWitness,
-  genesisCWitness,
 )
 where
 
@@ -26,14 +22,8 @@ import Cardano.Ledger.Binary (DecCBOR, EncCBOR, FromCBOR, ToCBOR)
 import Cardano.Ledger.Core.Era (Era (EraCrypto))
 import Cardano.Ledger.Credential (Credential (..), StakeCredential)
 import Cardano.Ledger.Hashes (ScriptHash)
-import Cardano.Ledger.Keys (
-  Hash,
-  KeyHash (..),
-  KeyRole (..),
-  VerKeyVRF,
-  asWitness,
- )
-import Cardano.Ledger.PoolParams
+import Cardano.Ledger.Keys (KeyHash (..), KeyRole (..), asWitness)
+import Cardano.Ledger.PoolParams (PoolParams (ppId))
 import Cardano.Ledger.Slot (EpochNo (..))
 import Control.DeepSeq (NFData (..), rwhnf)
 import Data.Kind (Type)
@@ -65,22 +55,10 @@ class
 
   getTxCertPool :: TxCert era -> Maybe (PoolCert (EraCrypto era))
 
-  mkTxCertGenesis :: ConstitutionalDelegCert (EraCrypto era) -> TxCert era
-
-  getTxCertGenesis :: TxCert era -> Maybe (ConstitutionalDelegCert (EraCrypto era))
-
 pattern TxCertPool :: EraTxCert era => PoolCert (EraCrypto era) -> TxCert era
 pattern TxCertPool d <- (getTxCertPool -> Just d)
   where
     TxCertPool d = mkTxCertPool d
-
-pattern TxCertGenesis ::
-  EraTxCert era =>
-  ConstitutionalDelegCert (EraCrypto era) ->
-  TxCert era
-pattern TxCertGenesis d <- (getTxCertGenesis -> Just d)
-  where
-    TxCertGenesis d = mkTxCertGenesis d
 
 -- | The delegation of one stake key to another.
 data Delegation c = Delegation
@@ -107,19 +85,6 @@ instance NoThunks (PoolCert c)
 instance NFData (PoolCert c) where
   rnf = rwhnf
 
--- | Constitutional key delegation certificate
-data ConstitutionalDelegCert c
-  = ConstitutionalDelegCert
-      !(KeyHash 'Genesis c)
-      !(KeyHash 'GenesisDelegate c)
-      !(Hash c (VerKeyVRF c))
-  deriving (Show, Generic, Eq)
-
-instance NoThunks (ConstitutionalDelegCert c)
-
-instance NFData (ConstitutionalDelegCert c) where
-  rnf = rwhnf
-
 poolCertKeyHashWitness :: PoolCert c -> KeyHash 'Witness c
 poolCertKeyHashWitness = \case
   RegPool poolParams -> asWitness $ ppId poolParams
@@ -128,9 +93,3 @@ poolCertKeyHashWitness = \case
 poolCWitness :: PoolCert c -> Credential 'StakePool c
 poolCWitness (RegPool pool) = KeyHashObj $ ppId pool
 poolCWitness (RetirePool k _) = KeyHashObj k
-
-genesisKeyHashWitness :: ConstitutionalDelegCert c -> KeyHash 'Witness c
-genesisKeyHashWitness (ConstitutionalDelegCert gk _ _) = asWitness gk
-
-genesisCWitness :: ConstitutionalDelegCert c -> KeyHash 'Genesis c
-genesisCWitness (ConstitutionalDelegCert gk _ _) = gk