+++ libs/cardano-ledger-test/src/Test/Cardano/Ledger/Generic/PrettyCore.hs
@@ -1399,13 +1399,13 @@ instance c ~ EraCrypto era => PrettyC (PoolCert c) era where prettyC _ = pcPoolC
 pcShelleyTxCert :: ShelleyTxCert c -> PDoc
 pcShelleyTxCert (ShelleyTxCertDelegCert x) = pcDelegCert x
 pcShelleyTxCert (ShelleyTxCertPool x) = pcPoolCert x
-pcShelleyTxCert (ShelleyTxCertGenesis _) = ppString "GenesisCert"
+pcShelleyTxCert (ShelleyTxCertGenesisDeleg _) = ppString "GenesisCert"
 pcShelleyTxCert (ShelleyTxCertMir _) = ppString "MirCert"
 
 pcConwayTxCert :: ConwayTxCert c -> PDoc
 pcConwayTxCert (ConwayTxCertDeleg dc) = prettyA dc
 pcConwayTxCert (ConwayTxCertPool poolc) = pcPoolCert poolc
-pcConwayTxCert (ConwayTxCertConstitutional _) = ppString "GenesisCert"
+pcConwayTxCert (ConwayTxCertCommittee _) = ppString "ConwayTxCertCommittee"
 
 instance c ~ EraCrypto era => PrettyC (ShelleyTxCert c) era where prettyC _ = pcShelleyTxCert